// whirlpool/smartcsb2c/Custom-apps/BE/worldpayconnector/node/typings/configs.d.ts
export interface configs {
  wpUrl: string,
  merchantCode: string,
  installationID: string,
  wpusername: string,
  wppassword: string,
  paymentAppName: string,
  wpVersion: string,
  xipayUrl:string,
  xipayUsername: string,
  xipayPassword: string,
  xipayMerchantid: string,
  sapmid: string,
  sapgl: string,
  xipayAmericanXpress: string,
  xipayDiscover: string,
  xipayMaestro: string,
  xipayMasterpass: string,
  xipayVisa: string,
  xipayVisaCredit: string,
  xipayVisaDebit: string,
  xipayVisaCommercialCredit: string,
  xipayVisaCommercialDebit: string,
  xipayVisaElectron: string,
  xipayECMC: string,
  xipayECMCCredit: string,
  xipayECMCDebit: string,
  xipayECMCCommercialCredit: string,
  xipayECMCCommercialDebit: string
}